pyproject.toml
[tool.poetry]
name = "cryptopals"
version = "0.1.0"
description = ""
authors = ["azuline <azuline@riseup.net>"]

[tool.poetry.dependencies]
python = "^3.7"
pycryptodome = "^3.9"
flask = "^1.1"
requests = "^2.22"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
isort = "^4.3"

[tool.black]
line-length = 89
target-version = ['py37']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
